// src/app/common/model/customerAddress.model.ts
export interface CustomerAddressSchema {
  _id: string;
  alter_phone: string;
  is_default: boolean;
  is_deleted: boolean;
  customer_id: string;
  name: string;
  phone: string;
  pin: string;
  address: string;
  city: string;
  state: string;
  type: string;
  createdAt: string;
  updatedAt: string;
}